repack-pdf: Allow not removing the temporary dir.

+++ repack-pdf.sh
@@ -1,6 +1,20 @@
 #!/bin/sh
 
-set -e
+#set -e
+
+KEEP_TEMP_DIR=0
+
+case $1 in
+    --help)
+        echo "$0 [--keep-temp-dir] <input.pdf>"
+        exit 0
+        ;;
+    --keep-temp-dir)
+        KEEP_TEMP_DIR=1
+        shift
+        ;;
+esac
+
 
 # Note: mktemp instead of the other insecure alternatives
 WORKDIR="$(mktemp -d)"
@@ -22,4 +36,7 @@ pingo -lossless -s9 -verbose=3 *.png
 img2pdf --verbose -o "$NEWNAME" *
 
 cd "$CURDIR"
-rm -rf "$WORKDIR"
+
+if [ $KEEP_TEMP_DIR = 0 ]; then
+    rm -rf "$WORKDIR"
+fi
